Improved config loader
