+++ src/service/ConfigLoader.ts
@@ -16,11 +16,11 @@
 import { existsSync } from 'fs';
 import * as _ from 'lodash';
 import { join } from 'path';
-import { Account, NetworkType } from 'symbol-sdk';
+import { Account, Address, MosaicId, MosaicNonce, NetworkType } from 'symbol-sdk';
 import { LogType } from '../logger';
 import Logger from '../logger/Logger';
 import LoggerFactory from '../logger/LoggerFactory';
-import { Addresses, ConfigAccount, ConfigPreset, NodePreset } from '../model';
+import { Addresses, ConfigAccount, ConfigPreset, MosaicAccounts, NodeAccount, NodePreset } from '../model';
 import { BootstrapUtils, Migration } from './BootstrapUtils';
 import { Preset } from './ConfigService';
 
@@ -29,7 +29,146 @@ const logger: Logger = LoggerFactory.getLogger(LogType.System);
 export class ConfigLoader {
     private static presetInfoLogged = false;
 
-    public static createPresetData(
+    public async generateRandomConfiguration(presetData: ConfigPreset): Promise<Addresses> {
+        const networkType = presetData.networkType;
+        const addresses: Addresses = {
+            version: this.getAddressesMigration(presetData.networkType).length + 1,
+            networkType: networkType,
+            nemesisGenerationHashSeed:
+                presetData.nemesisGenerationHashSeed || Account.generateNewAccount(networkType).publicAccount.publicKey,
+        };
+
+        if (presetData.nodes) {
+            addresses.nodes = await this.generateNodeAccounts(presetData, networkType);
+        }
+
+        const sinkAddress = Account.generateNewAccount(networkType).address.plain();
+
+        if (!presetData.harvestNetworkFeeSinkAddress) {
+            presetData.harvestNetworkFeeSinkAddress = sinkAddress;
+        }
+        if (!presetData.mosaicRentalFeeSinkAddress) {
+            presetData.mosaicRentalFeeSinkAddress = sinkAddress;
+        }
+        if (!presetData.namespaceRentalFeeSinkAddress) {
+            presetData.namespaceRentalFeeSinkAddress = sinkAddress;
+        }
+
+        presetData.networkIdentifier = BootstrapUtils.getNetworkIdentifier(presetData.networkType);
+        presetData.networkName = BootstrapUtils.getNetworkName(presetData.networkType);
+        if (!presetData.nemesisGenerationHashSeed) {
+            presetData.nemesisGenerationHashSeed = addresses.nemesisGenerationHashSeed;
+        }
+
+        if (presetData.nemesis) {
+            addresses.nemesisSigner = ConfigLoader.toConfig(this.generateAccount(networkType, presetData.nemesis.nemesisSignerPrivateKey));
+            if (!presetData.nemesis.nemesisSignerPrivateKey && addresses.nemesisSigner) {
+                presetData.nemesis.nemesisSignerPrivateKey = addresses.nemesisSigner.privateKey;
+            }
+        }
+
+        if (!presetData.nemesisSignerPublicKey && addresses.nemesisSigner) {
+            presetData.nemesisSignerPublicKey = addresses.nemesisSigner.publicKey;
+        }
+
+        const nemesisSignerAddress = Address.createFromPublicKey(presetData.nemesisSignerPublicKey, networkType);
+
+        if (!presetData.currencyMosaicId)
+            presetData.currencyMosaicId = BootstrapUtils.toHex(
+                MosaicId.createFromNonce(MosaicNonce.createFromNumber(0), nemesisSignerAddress).toHex(),
+            );
+        if (!presetData.harvestingMosaicId) {
+            if (!presetData.nemesis) {
+                throw new Error('nemesis must be defined!');
+            }
+            if (presetData.nemesis.mosaics && presetData.nemesis.mosaics.length > 1) {
+                presetData.harvestingMosaicId = BootstrapUtils.toHex(
+                    MosaicId.createFromNonce(MosaicNonce.createFromNumber(1), nemesisSignerAddress).toHex(),
+                );
+            } else {
+                presetData.harvestingMosaicId = presetData.currencyMosaicId;
+            }
+        }
+
+        if (presetData.nemesis) {
+            if (presetData.nemesis.mosaics) {
+                const mosaics: MosaicAccounts[] = [];
+                presetData.nemesis.mosaics.forEach((m, index) => {
+                    const accounts = this.generateAddresses(networkType, m.accounts);
+                    mosaics.push({
+                        id: index ? presetData.currencyMosaicId : presetData.harvestingMosaicId,
+                        name: m.name,
+                        type: index ? 'harvest' : 'currency',
+                        accounts,
+                    });
+                });
+
+                presetData.nemesis.mosaics.forEach((m, index) => {
+                    const accounts = mosaics[index].accounts;
+                    if (!m.currencyDistributions) {
+                        const caNodes = (addresses.nodes || []).filter((node) => node.ca);
+                        const totalAccounts = (m.accounts || 0) + caNodes.length;
+                        const amountPerAccount = Math.floor(m.supply / totalAccounts);
+                        m.currencyDistributions = [
+                            ...accounts.map((a) => ({ address: a.address, amount: amountPerAccount })),
+                            ...caNodes.map((n) => ({ address: n.ca!.address, amount: amountPerAccount })),
+                        ];
+                        if (m.currencyDistributions.length)
+                            m.currencyDistributions[0].amount += m.supply - totalAccounts * amountPerAccount;
+                    }
+                    const supplied = m.currencyDistributions.map((d) => d.amount).reduce((a, b) => a + b, 0);
+                    if (m.supply != supplied) {
+                        throw new Error(`Invalid nemgen total supplied value, expected ${m.supply} but total is ${supplied}`);
+                    }
+                });
+                addresses.mosaics = mosaics;
+            }
+        }
+
+        return addresses;
+    }
+
+    public generateAddresses(networkType: NetworkType, size: number): ConfigAccount[] {
+        return ConfigLoader.getArray(size).map(() => ConfigLoader.toConfig(Account.generateNewAccount(networkType)));
+    }
+
+    public generateAccount(networkType: NetworkType, privateKey: string | undefined): Account {
+        return privateKey ? Account.createFromPrivateKey(privateKey, networkType) : Account.generateNewAccount(networkType);
+    }
+
+    public generateNodeAccount(presetData: ConfigPreset, index: number, nodePreset: NodePreset, networkType: NetworkType): NodeAccount {
+        const name = nodePreset.name || `node-${index}`;
+        const ca = ConfigLoader.toConfig(this.generateAccount(networkType, nodePreset.caPrivateKey));
+        const node = ConfigLoader.toConfig(this.generateAccount(networkType, nodePreset.nodePrivateKey));
+
+        const friendlyName = nodePreset.friendlyName || ca.publicKey.substr(0, 7);
+        const nodeAccount: NodeAccount = { name, friendlyName, roles: nodePreset.roles, ca, node };
+
+        const useRemoteAccount = nodePreset.nodeUseRemoteAccount || presetData.nodeUseRemoteAccount;
+
+        if (useRemoteAccount && (nodePreset.harvesting || nodePreset.voting))
+            nodeAccount.harvesterSigning = ConfigLoader.toConfig(this.generateAccount(networkType, nodePreset.harvesterSigningPrivateKey));
+
+        if (!useRemoteAccount && (nodePreset.harvesting || nodePreset.voting))
+            nodeAccount.harvesterSigning = ConfigLoader.toConfig(
+                this.generateAccount(networkType, nodePreset.harvesterSigningPrivateKey || ca.privateKey),
+            );
+
+        if (nodePreset.voting) nodeAccount.voting = ConfigLoader.toConfig(this.generateAccount(networkType, nodePreset.votingPrivateKey));
+        if (nodePreset.harvesting) nodeAccount.vrf = ConfigLoader.toConfig(this.generateAccount(networkType, nodePreset.vrfPrivateKey));
+
+        return nodeAccount;
+    }
+
+    public async generateNodeAccounts(presetData: ConfigPreset, networkType: NetworkType): Promise<NodeAccount[]> {
+        return Promise.all(presetData.nodes!.map((node, index) => this.generateNodeAccount(presetData, index, node, networkType)));
+    }
+
+    private static getArray(size: number): number[] {
+        return [...Array(size).keys()];
+    }
+
+    public createPresetData(
         root: string,
         preset: Preset,
         assembly: string | undefined,
@@ -63,7 +202,7 @@ export class ConfigLoader {
         return this.expandRepeat(presetDataWithDynamicDefaults);
     }
 
-    public static dynamicDefaultNodeConfiguration(nodes?: NodePreset[]): NodePreset[] {
+    public dynamicDefaultNodeConfiguration(nodes?: NodePreset[]): NodePreset[] {
         return _.map(nodes || [], (node) => {
             const roles = this.resolveRoles(node);
             if (node.harvesting && node.api) {
@@ -116,7 +255,7 @@ export class ConfigLoader {
             throw new Error('A node must have at least one harvesting: true or api: true');
         });
     }
-    private static resolveRoles(nodePreset: NodePreset): string {
+    private resolveRoles(nodePreset: NodePreset): string {
         if (nodePreset.roles) {
             return nodePreset.roles;
         }
@@ -140,7 +279,7 @@ export class ConfigLoader {
         };
     }
 
-    public static expandRepeat(presetData: ConfigPreset): ConfigPreset {
+    public expandRepeat(presetData: ConfigPreset): ConfigPreset {
         return {
             ...presetData,
             databases: this.expandServicesRepeat(presetData, presetData.databases || []),
@@ -151,7 +290,7 @@ export class ConfigLoader {
     }
 
     // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-    public static applyValueTemplate(context: any, value: any): any {
+    public applyValueTemplate(context: any, value: any): any {
         if (!value) {
             return value;
         }
@@ -170,7 +309,7 @@ export class ConfigLoader {
     }
 
     // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-    public static expandServicesRepeat(context: any, services: any[]): any[] {
+    public expandServicesRepeat(context: any, services: any[]): any[] {
         return _.flatMap(services || [], (service) => {
             if (service.repeat === 0) {
                 return [];
@@ -188,7 +327,7 @@ export class ConfigLoader {
         });
     }
 
-    public static loadExistingPresetDataIfPreset(target: string): ConfigPreset | undefined {
+    public loadExistingPresetDataIfPreset(target: string): ConfigPreset | undefined {
         const generatedPresetLocation = this.getGeneratedPresetLocation(target);
         if (existsSync(generatedPresetLocation)) {
             return BootstrapUtils.loadYaml(generatedPresetLocation);
@@ -196,7 +335,7 @@ export class ConfigLoader {
         return undefined;
     }
 
-    public static loadExistingPresetData(target: string): ConfigPreset {
+    public loadExistingPresetData(target: string): ConfigPreset {
         const presetData = this.loadExistingPresetDataIfPreset(target);
         if (!presetData) {
             throw new Error(`The file ${this.getGeneratedPresetLocation(target)} doesn't exist. Have you executed the 'config' command?`);
@@ -204,22 +343,22 @@ export class ConfigLoader {
         return presetData;
     }
 
-    public static loadExistingAddressesIfPreset(target: string): Addresses | undefined {
+    public loadExistingAddressesIfPreset(target: string): Addresses | undefined {
         const generatedAddressLocation = this.getGeneratedAddressLocation(target);
         if (existsSync(generatedAddressLocation)) {
-            const presetData = ConfigLoader.loadExistingPresetData(target);
+            const presetData = this.loadExistingPresetData(target);
             return this.migrateAddresses(BootstrapUtils.loadYaml(generatedAddressLocation), presetData.networkType);
         }
         return undefined;
     }
 
     // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-    public static migrateAddresses(addresses: any, networkType: NetworkType | undefined): Addresses {
+    public migrateAddresses(addresses: any, networkType: NetworkType | undefined): Addresses {
         const migrations = this.getAddressesMigration(networkType);
         return BootstrapUtils.migrate('addresses.yml', addresses, migrations);
     }
 
-    public static getAddressesMigration(networkType: NetworkType | undefined): Migration[] {
+    public getAddressesMigration(networkType: NetworkType | undefined): Migration[] {
         return [
             {
                 description: 'Key names migration',
@@ -244,7 +383,7 @@ export class ConfigLoader {
         ];
     }
 
-    public static loadExistingAddresses(target: string): Addresses {
+    public loadExistingAddresses(target: string): Addresses {
         const addresses = this.loadExistingAddressesIfPreset(target);
         if (!addresses) {
             throw new Error(`The file ${this.getGeneratedAddressLocation(target)} doesn't exist. Have you executed the 'config' command?`);
@@ -252,11 +391,11 @@ export class ConfigLoader {
         return addresses;
     }
 
-    public static getGeneratedPresetLocation(target: string): string {
+    public getGeneratedPresetLocation(target: string): string {
         return join(target, 'preset.yml');
     }
 
-    public static getGeneratedAddressLocation(target: string): string {
+    public getGeneratedAddressLocation(target: string): string {
         return join(target, 'addresses.yml');
     }
 }
